Create a new class to contain methods that filter lists of acoustic Events
Issue #370
